agent-richy-v2/backend/requirements.txt
# Agent Richy v2 — Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.10.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
httpx>=0.26.0
sse-starlette>=1.8.0
